[Refactor] AppNewUpdateView  하위 뷰로 분리
- 앱 새로운 소식 영역 하위 뷰로 분리
